organize module json files
