+++ src/media/modules/module_10.json
@@ -1,0 +1,71 @@
+{
+    "perceptrons": {
+        "number": 10,
+        "title": "Module 10: Introduction to Machine Learning: Perceptrons",
+        "sections": [
+          {
+            "title": "Section 1",
+            "colorScheme": "dark",
+            "subsections": [
+              {
+                "title": "Subsection 1",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              },
+              {
+                "title": "Subsection 2",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              },
+              {
+                "title": "Subsection 3",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              }
+            ]
+          },
+          {
+            "title": "Section 2",
+            "colorScheme": "light",
+            "subsections": [
+              {
+                "title": "Subsection 1",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              },
+              {
+                "title": "Subsection 2",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              },
+              {
+                "title": "Subsection 3",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              }
+            ]
+          },
+          {
+            "title": "Section 3",
+            "colorScheme": "dark",
+            "subsections": [
+              {
+                "title": "Subsection 1",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              },
+              {
+                "title": "Subsection 2",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              },
+              {
+                "title": "Subsection 3",
+                "body": "lorem ipsum",
+                "imgSrc": "mod10.png"
+              }
+            ]
+          }
+        ]
+      }
+}
